Do not set empty disk serial number
Setting empty disk serial results in appearance of <serial></serial>
element inside <disk> section. This element resides above <address>
element in the XML-file. It was noticed, that this empty <serial>
element breaks the parsing of at least <address> element, which results
in uninitialized return value of Disk.getAddress().

Leave serial number string uninitialized in case SN was not set in the
source config.

Signed-off-by: Igor Bazhitov <ibazhitov@virtuozzo.com>