Identify Containers by their native UUIDs, rather than UUIDs, generated by a dispatcher
If Container is ideintifed by integer CTID, than dispatcher generates some UIID
for it for unification with VMs and internal (in-dispatcher) management. Low
level tools like vzctl or kernel are not aware of this generated UUID and thus
we have to convert UUID back to CTID for external for dispatcher tools. (#PSBM-34267)