bug #PSBM-39783
libvirt doesn't know about prl templates. thus we need to a) undefine
a libvirt domain on convertion to a template; b) define a libvirt
domain on reverse convertion. also skip libvirt application of
template configuration changes.

else if is redundant because the if about always returns.

igor@ doesn't like the else statement. ok