Do not set disconnected device state based on file path (PSBM-41676)
After we had moved disconnected device detection to
untranslatable_types, it is no longer needed to map empty disk image to
PVE::DeviceDisconnected when doing libvirt -> XML model translation.
Also, the presence of this mapping resulted in creation of a CDROM with
empty image path, which would be disconnected by default, whereas in VZ6
it was connected.

Signed-off-by: Igor Bazhitov <ibazhitov@virtuozzo.com>