Revert "ConfigCache do stat() on every get config action" (PSBM-42420)
This reverts commit 4ba4abd01cc90dfe519de09f83be2c269972964e.

During some actions, e.g. CT restore or disk resize (maybe others), the
VM config inside Dispatcher is updated several seconds after the
corresponding changes are made in the on-disk version of the config.
This time gap could reach one minute.
The source of these bugs was introduced in r1226040. There, the VM
config cache logic was changed to check config timestamps once in a
minute. Presumably, this was done to decrease the number of stats() on a
pstorage FS.