per countersf: read CT config from cache
It was found that Dispatcher reads config of each CT from disk while
collecting performance statistics. The config is required for retrieval
of count of vcpus.
Such behaviour increase load on Pstorage in case of large number of
active performance subscriptions.

Lets reuse Dispatchers cache and read the config from it.

#PSBM-41481