revert to snapshot with different vm configuration
Libvirt fails to revert to snapshot for a running domain if snapshot
has the config different from running one. But if option VIR_DOMAIN_SNAPSHOT_REVERT_FORCE
is applied it is possible. In this case domain will be stopped and
then started back again from snapshot. This option have no influence
if configs are compatible. Looks like it can be turned on for our usecases
unconditionally.

see #PSBM-43116

Signed-off-by: Nikolay Shirokovskiy <nshirokovskiy@virtuozzo.com>