make stack index and hint name correpondence consistent
* Problem

New vm is created in next steps:
1. create default config
2. convert config to libvirt and define domain
3. catch event from libvirt, convert libvirt config to dispatcher
and update dispatcher vm config.

On step 2 we use 'toBase26' function to convert stack index to
name hint. But on step 3 we use different mechanism and this
leads to libvirt and dispatcher config inconsitencies. This problem
occurs on config updates too.

* Solution

Use reverse function 'fromBase26' on step 3.