bug #PSBM-43948
libvirt sends shutdown event first and then stopped one when we stop
a vm gracefully. our handler produces stopped events on receipt of
both. thus we start handling the vm stop when it is still running
from the libvirt pov. we should switch a vm to stopped state only when
it is really stopped.