Check RAM size before enabling memory hot-plugging
When user enables memory hot-plugging for a VM the system applies
<MaxNumaRam> parameter that is not enabled if hot-plugging is off.
So, in the boot time the OS thinks it has at least 8GB of available
memory and configures it's buffers and processes based on 8GB
of available memory (this amount of memory can not be reclaimed).
And only after that balloon driver tries to increase it's size and
limit the VM memory by the configured memory size (<RAM>, --memsize).
So, if OS buffer size is more than --memsize the VM hangs
because balloon driver can not reclaim memory.

We assume that most OSes not require more than 1GB of RAM at boot time
in most cases. Because of that lets prohibit enabling of memory
hot-plugging support in a VM if RAM size is lesser than 1Gb.

PSBM-43554