fix config on migration
Virtuozzo have an option to change VM name and bundle destination
on migration. Libvirt supports changing config on migration too
we only need to pass the config on target as migration parameter.

Let's pass the config to the target and update it there and
then pass it back to the source on the check requrest - reply
stage. Then use the updated config as migration parameter.

Changing name needs another migration option - destination name.

see #PSBM-43764

Signed-off-by: Nikolay Shirokovskiy <nshirokovskiy@virtuozzo.com>