bug #PSBM-44822
we have 2 scenarios for creation/deletion of a vm: a prl sdk client requests
vm operation or an event about vm appearance/disappearance comes from libvirt.
in the formet case a dispatcher task drives necessary infrastructure changes.
in the later one these updates need to be performed in place.
also we may need to hide vm from dispatcher client without actual destroying.
for that purpose we have to pretend that the vm model is unavailable from the
registry and should be able to revert changes back.

define a vm if it is undeclared already is also an error