bug #PSBM-43935 aka Organize a secure tunnel for disks & memory data
qemu tunnel migration is broken and they don't want to fix it. they
provide only unsecure unmanaged online migration between 2 peer qemu
processes. there is an option to tell to the target qemu which ports
it should listen for incomming connections from the source.
let's cheat it the same way as we do for libvirt traffic and tunnel
both state and nbd data through the dispatcher securred connection
for the sake of encyption.