Save VM state in the ClusterOptions part of the VM config (PSBM-45400)
When shaman relocates a VM from a crashed node, it needs to know the
state of the VM at the moment when the node had crashed. If the VM was
running - shaman must start this VM upon relocating it to another node.
In VZ6 we've used Settings/ClusterOptions/Running flag to save the state
of a VM and to retrieve it in shaman when relocating the VM.
Note, that setting the "Running" flag to 0 when a VM transitions to
the paused state is consistent with VZ6 behavior.

Signed-off-by: Igor Bazhitov <ibazhitov@virtuozzo.com>