bug #PSBM-45551
we always update vm state when a defined event comes. that update may
clash with one for a state event. state events carry correct sequence
of state transitions but they lack time marks thus we cannot be sure
that we place manual state update into correct place in the sequence.
still we need to start with some state when a new vm is discovered.
let's pull vm state on reconnect to libvirt and on new vm define.