bug #PSBM-45715
libvirt doesn't issue a state change event in case a vm being in
stopped state is reverted from snapshot to the same state. we sent
event from the pull worker to workaround. revert command handler
waits for an event though.  we don't send them anymore for existing
vms thus it hangs.
let's implement a dedicated pull worker for switch to snapshot case
which sets the vm state for -> stopped transitions