Unregister VMs that are already registered on other nodes in the cluster (PSBM-45516)
Consider the following situation.
This node is registered in the cluster and had just crashed. shaman had
relocated a VM from this node and registered it on another node in the
cluster. After that, this node had been rebooted and is back online.
Here is what we have on Dispatcher start:
- The VM is registered both in libvirt and Dispatcher.
- Server UUID of the VM differs from our server UUID.

In this case we need to automatically unregister the VM from this node.

Signed-off-by: Igor Bazhitov <ibazhitov@virtuozzo.com>