Distinguish attached backups from ordinary disks in guest by SN (PSBM-44137)
In VZ6 we have distinguished the attached backups from ordinary disks
using a well known prefix in the HDD's model, e.g. ID_MODEL=__bckp1.
In VZ7 we can't use the HDD model string because it's value is hardcoded
in QEMU. Instead, we'll use the HDD's serial number which we are manually
generating inside the Dispatcher.

Signed-off-by: Igor Bazhitov <ibazhitov@virtuozzo.com>