Enhance checking for IP conflicts. #PSBM-33726
Currently, the IP conflicts are checked between VMs only.
This fix adds IP conflicts checking when IP address is set for CT, and it takes existing CT IP addresses into account when checking for a conflict.
If auto-applying of the configuration for an adapter is disabled (--configure no), its IP address is not considered when checking for duplicates.

Signed-off-by: Anton Sanarov <asanarov@virtuozzo.com>