Incremental virtual machines backup & proper suffix
Implemented incremental backups in new 'push' scheme. Difference from the
previous model:
* Calculate final archive name (including sequence number) from the beginning.
  In previus version Acronis libraries did this 'magic', which is not good, but
  we had to follow them.
* Calculate current PIT and previous one. These PITs are unique string bitmask
  identifiers, which are propagated to qemu and preserved by it. Next time we
  give previous bitmask ID to qemu and thus retrieve only changed blocks from
  that PIT.

As a side consequence of this change, declare correct archive suffix finally.
This change was supposed to be done long ago, but now it has a strong reason to
go in - necessity to add sequence number to it.

ToDo: handle unexpected bitmask absence (qemu crash or hard stop)

PSBM-44441