bug #PSBM-44755 aka migration hangs in dispatcher if libvirt connection keepalive triggers
we deadlock when the libvirt tunnel socket is open but the libvirt
task cancellation is requested because the libvirt client stub waits
for a message from the socket which closing in turn is blocked by a
wait for the libvirt thread completion.

the tunnel would be terminated gracefully if we postpone the libvirt
thread waiting. let's move that wait outside of the move join state
to execute it a bit later.