Fixed backup of stopped virtual machines
In new "push" scheme we ask qemu process to do a job. There is no qemu process
for stopped VMs - so, let's start it.

This patch starts a qemu process in guest paused state, when there are no vcpus
activated and guest is not booted. On stop, we check that VM is still in this
state - and stop it if so, returning it to the original state.

PSBM-44441