Close i/o devices when AuxChannel detects error or eof #PSBM-44027
In order for Vm exec to be finised when guest domain was stopped,
we need to terminate Vm::processStd(). This normally should be
done by setting eof condition on stdout/stderr ReadDevice.

We set eof from device close() which in turn will be called from
AuxChannel close(). Notice: m_lock should be unlocked to avoid
deadlock with removeIoChannels() and AuxChannel should be in
sOpen() == false condition before calling device close().