Directly pass ploop device instead of VD wrapper to VM config (PSBM-44137)
In VZ6, we needed a VirtualDisk wrapper over a ploop device in order for
this ploop device to be accessible as a HDD inside a VM.
In VZ7 we no longer need such wrapper since libvirt/QEMU supports block
devices as a virtual HDD data storage.
Setup a symlink to a ploop device when a backup is attached to a VM and
use this symlink as a virtual HDD backend.

Signed-off-by: Igor Bazhitov <ibazhitov@virtuozzo.com>