bug #PSBM-47322 aka Error occurred while edit VM with code [0x80000105][PRL_ERR_VM_UUID_NOT_FOUND]
edit vm was not ready to convert a template back into a vm. fi it
did't search inside the templates directory for an item - actual
reason of the failure.
take into account all the directories available to the client when
searching for a directory item inside edit vm task. then use the
item ident instead of the client default directory uuid.