Fixed restoration of Containers and VMs if it takes too long
If restoration is longer than 10 minutes, the task is cancelled due to default
timeout on disp-to-disp communications.  Another problem is that callback is
lengthy and blocks receive of packages from the other side.

This patch fixes above deficiencies, making a callback on the source side
asynchronous (detach image restoration) and customizing a timeout on the target
side (giving a space for cancellation as bonus).

PSBM-47652