set the memory statistics collection period for running vms only
libvirt api is an example of a bad api design. they have a handle
to tune the memory statistics collection period which is targeted
for running domains only. thus it has no effect if we use it for
domains in other states.
fortunatelly, we are notified when a vm is being switched to the
running state. let's pull the handle at that moment.