Active bitmap detection mechanism
We need this mechanism in order to know in advance if a requested bitmap is
absent in qemu/ploop.  In this case we should force a full backup creation.

This is important part of new backup scheme - because a probability to have no
expected bitmap in a running ploop/qemu is significant. A bitmap could be lost
due to node's crash, forceful VM/CT stop or restoration from a backup.

PSBM-47796