bug #PSBM-48163 aka PlsCtlError: prlctl returned 255
we have a race between a defined event which comes after a new
config is applied and a moment when a device becomes attached
to a running vm. libvirt doesn't report block info when a devices
is in persistent config only thus we rewrite the disk size value
with 0.
transfer the size value from the old config if 0 is in the new
one for a hard disk device.