bug #PSBM-48602 aka dump: core-Libvirt::Host-1466302973.24988
code inside registerExclusiveVmOperation has a window without lock.
it expects that m_opHash is updated inside that window. we have 2
qt mutable iterators in that scope which are destroyed on exit which
is after the window. they alter their containers from the destructors.
the containers may be destroyed already which results in a ub.

limit life time of the qt iterators to a scope where the lock is
held.