bug #PSBM-48229 aka Pipe empty or closed
we don't wait for the stdin completion stage but still wake up
a condition. that wake up is treated as a signal that the fin
stage is completed and a task sends a result package which arrives
simultaneously with stdout/stderr event at a client. because of
the race we may not deliver content from the event to the client.

don't wake up stage from the stdin close callback.