bug #PSBM-48823 aka Template switched back to VM could not be started
we registered all the entries from all the vm directories as
listeners for defined events including templates which is wrong.
the registration info contains the vm directory uuid. during a switch
the vm moves into another directory invalidating the booking data.
when a defined event arrives we cannot find the vm using booking
info.
need to skip everything which is not a template when a connection
to libvirt is established. also we should not update directory
items when the template status changes.