bug #PSBM-48409 aka testStartStop TIMEDOUT: Function wait timeouted
exec task wait for a guest process exit and for completion of streams.
we have a process exit and guest agent restart thus no eof events
for the streams.
implement a naive approach for detection of the protocol resets. we
remember when the execution finishes and give another spin in timeout
handler to complete the streams stuff.