Encrypted VNC for vms
we have a feature in pcs6 to setup encrypted connections to VNC
servers of vms. in vz7 it is qemu who manages the VNC. it does
this on start of a vm.  we don't need to launch a separate
process for a VNC server anymore. still we have to start an
stunnel to encrypt the VNC traffic. also we have to setup a vm
VNC server to be accessible from the localhost only.
well, let's start a background job when a vm start is detected
when the dispatcher is configured to encrypt VNC traffic. that
job responsibility is to launch an stunnel instance and to
register a monitor for the stunnel.
also we don't have to override the encryption enabled flag in
vm config and use a runtime port value for unencrypted version
only.