bug #PSBM-48772 aka Start local VM during restore on dst should be fail
start is unaware of exclusive locks. thus it proceedes ignoring a
conflicting operation running simultaneously.

the proposed solution suggests to grab the lock for the start task
then do actual start of a vm and grab another lock (an indicator one)
the start task lock is released in the end of the operation but the
indicator lock stays until the vm is deactivated which is detected
by its state machine.