Apply "raw" CT parameters when applying new CT config. #PSBM-39266
Some container parameters which aren't used and parsed in the dispatcher
but may be used at higher levels age placed in the config as raw parameters
and can be retrieved and changed by getRawParam/setRawParam. Unfortunately,
applying these changes back to the container config has not been implemented.

This patch implements applying raw container parameters back to the container config.

Signed-off-by: Anton Sanarov <asanarov@virtuozzo.com>