implement a new tunnel naming schema for migration.
we used the local peer name for migration because it mapped 1 into 1
at the target side. this is not true for backups because the ports
that we listen locally differ from those at the target.
let a customization of a tunnel name by setting the channel property
of the source server socket object. the packer uses that information
to construct a package for the target.