introduce tunnels for backup/restore operations.
new backup engine doesn't support encryption of the intermediate
traffic in case of a remote operation.
let's use the same approach we did for migration i.e. provide
to a source an url to a socket being listened by the local
dispatcher. it receives all the traffic and transmits it to
the target via an encrypted inter-dispatcher channel. the target
dispatcher connects to the actual target side of the operation
and forward the traffic from the channel.