fix for assert in Qt
qt doesn't allow to notify receivers living in a thread that is
not the same as the sending one. there is an assert in the code
which is a nop in release builds that is why it doesn't show up
still.
fortunatelly, the problem event relates to a qobject destruction
which can be handled by a deleteLater event instead of a direct
deletion by the source thread. add an appropriate callback to
constructor of the qsharedpointer container for the qobject to
avoid the assert.