fix assert in qt
qt forbids receiving queued events in a thread which is not the one where a qobject lives.
restore task creates a qprocess instance implicitly in context of a disp connection thread
when a request to mount an archive images is being handled. the instance gets owned by that
thread. later qt asserts because of invalid thread affinity when finalyzeTask in the task
thread invokes indirectly an events related logic inside the qprocess.
we need to move creation of the qprocesses to the task thread to make qt happy. in order
convert mountImage into a slot and connect directly to a Target::IO instance residing in
the task thread. this way mount requests get handled inside the task thread.