Check CT state during encryption on libvzctl level (PSBM-52950)
Currently we explicitely prohibit all disk encryption operations on the
existing encrypted disks when a CT is running. But the encryption key
could actually be changed on a LUKS partiton when it is mounted, because
only a small amount of data in its header is substitued, while the
actual encrypted disk data left unchanged. This is done using
'cryptsetup luksChangeKey' command.

Move the CT state checks to the libvzctl (libploop) level, since
the cryptsetup tool is executed there.

Signed-off-by: Igor Bazhitov <ibazhitov@virtuozzo.com>