fix network type reconfiguration #PSBM-53573
Bond card mac and bridge are not required for host-only networks.
Given the fact that the user could leave these fields in a configuration
while he changing network's type (as prlsrvctl do), we must clear these
data. In other case Transponster puts mac and bridge name in the network xml.