bug #PSBM-54367 aka other_cases.second_restore TIMEDOUT: Function call_sdk_function timeouted
connection request comes when the tunnel is uninitialized. thus the
request is discarded and there is no way to attach to the nbd server.
we quit the event loop in the end of the file copy stage. it is
required for the old protocol only. the new one enters the event loop
again anyway.
don't quit the event loop in the end of the file copy stage for the
ndb based implementation.