bug #PSBM-54935 aka usb contoller is still in listing
we add an usb device on processing of a usb controller. the controller
is always there because libvirt adds it silently anyway. thus the
controller cannot be a marker. we need a usb device actually. we add
an input usb mouse device when usb devices are inside the dispatcher
version of the vm configuration.
let's leverage that fact and add an usb device to the dispatcher config
only when a usb input device is inside the libvirt config.