bug #PSBM-55528 aka w2k12 installer doesn't see VM's hdd
windows needs drivers for virtio devices. an fdd with drivers is
attached to the vm on creation from the default config. the device
model is created by sdk which requires from every user following a
certain sequence of actions. the api doesn't restrict users to
not perform these activities. thus not working configurations can
still be produced. there are no means in the api to fix the
incorrectness.
let's move the fdd device attachement into the dispatcher and perform
it as a reaction on signal about a new vm registration. emit that
signal as the very last step of the register vm task.