Move a firmware log file outside from a VM home directory #PSBM-54401
Qemu opens a log file in RW mode. This prevents from opening the file
from another server in case of the log file is placed on vstorage.
Qemu tries to open this file when VM is migrating that leads into error.

Let's move this log file outside form a VM home directory.