Fixed sending of updated VM configuration event on convertion to template.
The root cause of the issue is that in vz7 we store templates in another vm
directory. Thus all VM subscribers are silently lose it after conversion to
template - because there is no more such item in previous directory, where they
are subscribed to. The patch fixes this deficiency, sending CONFIG_CHANGED
event in this particular case (conversion to/from template).

Additional fixes:
* specify isTemplate property correctly in VmDirectoryItem (for correct listing
  if only identity info is requested)
* notify internal subscribers on VmConfig change in a particular overload of
  sendVmChangedEvent() function - it looks this call was simply missed there.

PSBM-56343