bug #PVA-35989 aka When "host cursor" leaves the area of the console, cursor stops moving, even in Full screen mode
libvirt adds a usb controller implicitly even if no controller was
in a define request. that controller is without the model attribute.
we ignore this kind of controller and treat the config as the one
without a usb controller configured.
libvirt uses piix3-uhci as a default model for the implicit USB
controller. bring that knowledge into the dispatcher code.