remove redundant locks in edit vm code
the first block contains duplicate calculation. it also holds the
dir manager lock for too long which blocks other users of the lock.
lock operations from the second block are performed when the lock
is already grabbed (the first operation of the enclosing block). thus
they are redundant. just meaningless calculations.