bug #PSBM-58437 aka prlctl resume failed due to timeout
resume is executed synchronously. the event loop which is in charge
of handling the timeout is not spining then. when the libvirt part
finishes we enter the loop. it is required to catch a transition
to the running state. when a libvirt operation takes more time than
the timeout value the timeout handler fires immediately after return
to the loop and sends the error.
there is no way to abort the libvirt call. it is suggested to not
restrict suspends with a timeout.