bug #PSBM-58567 aka prlct "--on-crash restart" vm option does not change into "paused" after 3 reboots
after a number of restarts we switch on crash mode of a problem vm
into the pause one. we use the very same function to update the
dispatcher xml model as we do for updates coming from the libvirt.
that function complements the xml model constructed from a libvirt
xml with the dispatcher internal data which is copied from the
previous version of the config. thus we overwrite changes when the
new config comes from dispatcher itself.
let's apply the untranslatable_types logic only when the home path
value of the function argument is empty which means we pull the
config from libvirt.