bug #PSBM-60527 aka SDK.PrlVmDev_ResizeImage with PRIF_DISK_INFO returns no data
restore info reporting by the resize image task in a way similar
to what we have in pcs6. prl_disk_tool dumps info into stdout plus
it writes warnings in stderr. parse that data to fill the
CDiskImageInfo data structure and send that structure to the client
when the image info is requested.