bug #PSBM-60193 aka deleting VM backup with --keep-chain failed
original implementation of backups coalescing was incorrect because
it corrupted ids of backups in the sequence. there is no restriction
on continuity of incremental numbers. also there is no limit on
format of id of the first backup in a sequence.
thus we introduce abstraction - backup sequence and make all
backup related operations use it. then backup coalescing becomes
trivial. we merge target backup with its direct next in the
sequence and update the metadata.