bug #PSBM-61858 aka migration is failed with 'The file .brand/dispatcher-build already exists on the destination node.'
CVmMigrateHelper::GetEntryListsVmHome collects all the file entries
in a vm private folder. branding files should not be transfered in
clone mode.
let's remove them explicitly from the list. the clone mode is
supported for vms in the stopped state thus there is nothing to
update in the running migration branch.