bug #PSBM-61470 aka Console doesn't work after successfully VM migration
config save in the end of a shared migration is not required anymore
because we do save config in a background thread unconditionally
processing an appropriate event from libvirt.
just remove it. also we should not load the vm config in the begining
of a migration because this leads to the config cache invalidation
and all the transient info gets lost, fi: websocket port for a
running vm. use cached version. it is good