bug #PSBM-62445 aka The folder "/vz/vmprivate" used to store the virtual machine files is not available.
there is a superfluous test for existance of the vmprivate folder.
its purpose is to workaround a case when the folder is on an
unreachable mount point. we don't have anything similar for containers
and there are no complains about the issue.
let's drop the test. just create the folder on demand.