bug #PSBM-62366 aka second migration in chain host1->migrate-> host2->migrate->host1 failed
merging snapshots in the end of a migration over pstorage results
in implicit change of persistent config of the vm. libvirt doesn't
notify its clients about the change. the dispatcher config of the
vm stays with unpatched shared disks paths while the files are not
there anymore.
libvirt team doesn't cooperate thus it is suggested to imitate a
receipt of a defined event after the merge.
event handling callback requires a model of libvirt vm list. that
object is unreachable for code outside of the libvirt liar. though
we have a hub for libvirt events. let's place a mock there with the
model. that mock is responsible for triggering a libvirt event
processing. then poke that mock after a shared disk merge in the
end of a migration.