Set SCSI bus by default for CT root disk
CT root disk has bus PMS_UNKNOWN_DEVICE, so we can't get correct
device name sd* or hd* for it in libvirt.
If we add a new disk via prlctl than it has SCSI bus by default.
It will be better if we have a specific bus type instead of
unknown value for root disk.