bug #PSBM-55015 aka Can't create backup VM in VZ7 if Node's hostname contains upper letters
there are several ways to setup a hostname for a linux host. from
the local view these hostnames would be accessible perfectly. qt
uses some method to extract the local hostname. it produces single
string. thus the direct comparison we use now is insufficient.
let's resolve the hostname which doesn't match into a qhostinfo
and then try to match the list of host addresses of the resolve
result and the list of addresses assigned to local network interfaces.