bug #PSBM-56656 aka backup ERROR: Details: Error: Block job is failed
the client report about an io problem may be obscure and it's
impossible to find exact reason of the failure. on success path
the inter-dispatcher connection is still active. any io error
also results in a disconnect finally. but, cancel disconnects
explicitly.
let's report a dedicated error in case the connection is down and
the task has not been cancelled.