bug #PSBM-58310 aka Not every VM starts on host boot if autostart is on
we have a limit on the number of dispatcher connections. that limit
doesn't take into account nature of a connection. thus it is applied
for all sources and users.
we use systemd units for the auto-start ves feature. for every
auto-start ve we generate a unit. these units start simultaneously and
connect to dispatcher. some connections fail if their number is above
the limit.
let's not apply the limit for local connections established from a
user with 0 uid.