bug #PSBM-59178 aka Race if attaching multiple backups simultaneously (backup was attached as regular disk without "backup" attribute)
we allow concurrent edit of a vm config. multi edit dispatcher protection
is not enough because it doesn't provide a critical section. we are
unable to detect conflicts between tentative configurations and test
each one against the current one only. then they are applied in an
unpredictable order if no conflicts were detected.
let's forbid simultaneous execution of several edit commit tasks and
rely on the exclusive operation lock to implement a critical section.