bug #PSBM-63944 aka prlctl returned error after trying to disconnect backup disk.
the problem is several tasks contend for the exclusive vm lock. and the prlctl
one quits without waiting. this is due to the resolution code leaves when
the condition is true initialy. but we get here because the conflict is detected
which means we cannot rely on the flag and need to wait unconditionally.
it's of no use to skip waiting when condition is true. always wait for a resolution.
also it makes sense to re-lock after an unsuccessful conflict resolution to
allow other lock waiters to make progress and may be resolve the conflict.