bug #PSBM-62516 aka Incorrectly constructed PRL_ERR_COPY_VM_INFO_FILE error message
error reporting design in dispatcher implies that the error can be
supplied with extra parameters. thus a code alone cannot be enough
for a client to construct complete error message. not all users of
CDspVmBrand api need details of a failure.
let's accept an optional dsp failure to fill details of an error
when a caller wants them.