bug #PSBM-54476 aka Enabling/disabling vnc encryption breaks vnc access to existent VEs
now we apply special VNC settings to a VM config if encryption for
VNC is enabled in dispatcher. the config is not updated when the
encyption gets disabled. a user sees something which he didn't
configure. this confuses.
we use stunnel to disguise the VNC server. but the tool always
encrypts. we suggest to always setup special VNC config for a VM
and tunnel the VNC traffic with the socat tool started in either
encrypted or unencrypted mode depending on the server settings.