bug #PSBM-65346 aka Failed to backup VE after removing full backup with --keep-chain
a base backup in a chain directory changes after coalescing the head
of the chain. access results in not found.
use general approach to access backup metadata for loading the ve
config from the base backup of a chain to determine if the disks
setup has changed since the last base backup.