bug #PSBM-65670 aka snapshot switch is stuck
dispatcher doesn't produce a state change event after a switch to
snapshot of the vm in paused state. we simply ignore a corresponding
libvirt event and hope to generate one when a defined event comes.
unfortunatelly, we do for a vm in stopped state only.
let's generate for a vm in paused state as well.