bug #PSBM-65819 aka other_cases.backup_several_vm_in_parallel FAIL: dump: core-{80c9b2b8-1cef--1493639631.3515
qt uses some obscure code to load and use libresolve to calculate
local domain name of a host. it is not thread safe and depends on
libresolve loading which may occur several times without an error
which results in overriding the value of a global variable not
protected with any lock.
we use the result of QHostInfo::localDomainName() in a single place.
let's protect the call with a mutex.