Add validation for max port number in manual mode (65535) PSBM-66051
It is possible to set value that exceeds maximum port number. This value in VM config will become 0 and the next attempt to set other settings or start this VM will fail.
This patch adds validation for max port number which equals to 65535 so there will be no possibility to assign the incorrect value.