bug #PSBM-66230 aka prlctl set {vmid} --device-disconnect doesn't work
update config that we perform as a part of a device disconnect
doesn't update state of the devices. the edit that prlctl sends
afterward still uses connected as the state value. the device
returns to connected state.

we have an example of updating a device state for optical disks when
a tray event is processed. let's generalize that experience to
set a device state when a device connected/disconnected event is
handled.