bug #PSBM-66121 aka Cannot connect to VNC server (in case 5900 port is selected)
1. we copy a qemu vnc port from the libvirt config for the vm which
is wrong now because the port listens on the loopback address only.
thus we need to start reporting only after the socat process starts.

2. every edit vm task resets the vnc ports for auto mode and we loose
actual values for vnc ports. need to preserve them.

3. stop of a vm doesn't remove the vnc ports from the vm config.
we report something unrelated until a next start of socats. need to
sweep the ports from the vm config.

prohibit a vm device connect/disconnect when the vm is not running

this is exactly what we have in pcs6