correct restore of the vnc server proxy after a qemu upgrade
a upgrade of qemu restarts the vnc server which leads to shutdown
of a socat proxy the dispatcher brings for the vnc port. this
also leads to reset of the vnc port value in the vm config to 0.
thus we cannot restore the port after upgrade because the old
value is lost and the socat won't start with 0 port.
have to use setup instead.