bug #PSBM-66709 aka VM state != VMS_STOPPED after stop operation (SDK API)
if timeout is expired during a stop of a vm we simply kill the vm
and return immediately. in the background processing of the state
transition ocurs but the client is allowed to proceed and may see the
vm still in running state which is not the case.
let's relax quit condition for the event loop and allow a timeout
handler to decide individually if the loop is to be terminated or
the command processing should proceede. introduce a signal and
establish connection with the loop quit slot.