bug #PSBM-66881 aka vzt-sata-hotplug-centos-7.3-x86_64-virtio-tools: XML error: target 'vdb' duplicated
2 problems. we have no order for processing events for a vm thus the
handling occurs in unpredictable order. a connect/disconnect device
event handler receives the device alias as a reference to the device.
we don't know the alias for a new device because it is generated by
libvirt when the device is being plugged in.
to address these problems the patch proposes to:
- serialize execution of event handlers for a vm;
- always patch devices aliases when a device connect/disconnect event
is being handled.