bug #PSBM-66950 aka vzt-pcs6upgrade-*: for all CDRom's attr Connected changed from 0 to 1
for cdrom's disconnect means tray open and empty image path in the
libvirt vm config.
add a case for empty systemName to set a device connected state to
disconnected.
also we quit registering of an exclusive operation too early in
presence of contention. it appears because we use wakeAll to wake up
conflict waiters instead of wakeOne (which was a remedy for another
issue). let's spin until the timeout instead of 2 loops.