bug #PSBM-66786 aka Degradation of total live migration time: 25% (Windows) / 50% (Linux)
we process a vm disappear synchroneously but other events are handled
in a separate thread. thus we have a race between the vm removal
during migration at the source side and removing its exclusive lock
in an auxiliary thread. the lock may stay and prevent grabbing new
one with the same command when the vm is being migrated back.
let's remove the race.