use another command for atomic edit exclusive lock
we use DspCmdDirVmEditCommit now for the purpose. that command conflicts
with many others which results in hangs for atomic edits.
DspCmdDirVmEditCommit relates to whole ve. in case of atomic edit
we are interested in update of a small amount of configuration in
response to some external activity. it is about the config not the
vm as a whole.
let's employ unused DspCmdCtlApplyVmConfig to ensure a fine grained
locking.